Remove syntastic and only use ALE for syntax checking
Note: this also makes it so that only the linters specified in vimrc for
ale will be used.  They must be explicitly named to be turned on.

This was mostly because ale was complaining about prettier errors on my
ebay project in ts files, since it was applying eslint rules to ts files
as well.