Make more than just the explicitly named linters run. Auto fix js files
When explicitly naming the linters to use, some errors were falling
through the cracks and not getting brought up. Letting ale use whatever
it wants to point out errors, even though there are still some annoying
prettier errors in the ts files.  (I'd love to get rid of those styling
mistakes we made, but we can't touch that code - fuck).

Since ALE is smart enough to know what corrector to use between ts/js,
we're using that now to automatically prettify the code.  There was an
issue where we'd save it, then ale would run and fix it, then it would
change the stuff, and we'd have unsaved changes.  Now we made an
autocommand to save it after ale has run so that it'll auto-save the
changes